requirements.txt
Jinja2==3.0.3
PyJWT==2.3.0
aiohttp==3.10.3
bcrypt==4.2.0
docx2txt==0.8
faiss-cpu==1.8.0.post1
langchain-community==0.2.11
langchain-core==0.2.28
markdown==3.0.0
openai==1.55.3
psycopg2==2.9.9
pymilvus==2.4.7
pypdf==4.3.1
pytest-cov==5.0.0
pytest==8.3.2
scikit-learn==1.5.1
tiktoken==0.7.0
unstructured==0.15.1
chromadb==0.5.11
chroma-hnswlib==0.7.6
markdown==3.0.0
gTTS==2.5.3
pdf2image==1.17.0